pyproject.toml
[project]
name = "healthymealsai"
version = "0.1.0"
description = "Healthy Meals AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
